requirements.txt
robotframework
robotframework-seleniumlibrary
selenium>=4.18
requests
